EN hunted down some warnings

+++ lib/CoreUtils/vector.hpp
@@ -248,7 +248,7 @@ public:
 
 	/// copy the elements to somthing designed after the output iterator model
 	template<class InputIterator> void copyFrom( InputIterator start, InputIterator end ) {
-		LOG_IF( std::distance( start, end ) > SIZE, Runtime, error )
+		LOG_IF( size_t( std::distance( start, end ) ) > SIZE, Runtime, error )
 				<< "Copying " << std::distance( start, end ) << " Elements into a vector of the size " << SIZE;
 		std::copy( start, end, CONTAINER::begin() );
 	}
